Moving the new test to its own file as requested in review
